+++ jest.config.js
@@ -24,6 +24,7 @@ module.exports = {
         'zoid/src': 'zoid',
         'jsx-pragmatic/src': 'jsx-pragmatic',
         '@paypal/sdk-client/src': '@paypal/sdk-client',
+        '@paypal/sdk-constants/src': '@paypal/sdk-constants',
         'belter/src': 'belter'
     },
     collectCoverageFrom: ['./packages/**'],