+++ packages/components/tests/unit/spec/modal/parts/Container.test.js
@@ -39,7 +39,10 @@ describe('<Container />', () => {
                     product
                 }
             }
-        ]
+        ],
+        meta: {
+            trackingDetails: {}
+        }
     });
 
     const content = (