+++ packages/library/src/zoid/message/component.js
@@ -169,13 +169,18 @@ export default getGlobalVariable('__paypal_credit_message__', () =>
                         const { account, merchantId, index, modal, getContainer } = props;
                         const { messageRequestId, displayedMessage, trackingDetails, offerType } = meta;
 
-                        logger.addMetaBuilder(() => {
+                        logger.addMetaBuilder(existingMeta => {
+                            // Remove potential existing meta info
+                            // Necessary because beaver-logger will not override an existing meta key if these values change
+                            // eslint-disable-next-line no-param-reassign
+                            delete existingMeta[index];
+
                             return {
                                 [index]: {
                                     messageRequestId,
                                     account: merchantId || account,
                                     displayedMessage,
-                                    ...trackingDetails
+                                    trackingDetails
                                 }
                             };
                         });