+++ packages/renderer/src/locale/US/GPL/mutations/gpl.js
@@ -1,12 +1,12 @@
-import Logo from '../logos';
+import Logo from '../../../../message/logos';
 import { xSmallFallback, textWrap, messageLogoWidth, altNoWrap, setLogoTop } from '../../../../message/mediaQueries';
-import { flexLogoMutations, textLogoMutations } from './common';
+import { flexLogoMutations, textLogoMutations } from '../../../../message/logoMutations';
 
 const flex = [
     [
         'default',
         {
-            logo: Logo.PRIMARY.WHITE,
+            logo: Logo.PP_PAYPAL.WHITE,
             headline: [
                 {
                     tag: 'xsmall'
@@ -69,6 +69,7 @@ const flex = [
             ]
         }
     ],
+    ['color:white-no-border', { logo: Logo.PP_PAYPAL.COLOR }],
     ...flexLogoMutations
 ];
 
@@ -84,7 +85,7 @@ export default {
                     messageLogoWidth(false, textSize * 4, textSize * 1.25),
                     setLogoTop(textSize * 20)
                 ],
-                logo: Logo.PRIMARY.COLOR,
+                logo: Logo.PP_PAYPAL.COLOR,
                 headline: [
                     {
                         tag: 'medium',
@@ -121,7 +122,7 @@ export default {
                     altNoWrap(textSize * 10.6),
                     messageLogoWidth(textSize * 1.75, textSize * 4, textSize * 1.25)
                 ],
-                logo: Logo.PRIMARY.COLOR[0]
+                logo: Logo.PP_PAYPAL.COLOR[0]
             })
         ],
         [
@@ -150,7 +151,7 @@ export default {
             'logo.type:inline',
             ({ textSize }) => ({
                 styles: [xSmallFallback(textSize * 18), `.message__logo { width: ${textSize * 4}px }`],
-                logo: Logo.INLINE.COLOR,
+                logo: Logo.NO_PP_MONOGRAM.COLOR,
                 headline: [
                     {
                         tag: 'medium.2',