+++ packages/renderer/src/locale/US/PPC/mutations/pala_single_eqz.js
@@ -1,5 +1,5 @@
 import Logo from '../logos';
-import { altContentMediaQuery, primaryContentMediaQuery, textWrap } from './mediaQueries';
+import { altContentMediaQuery, primaryContentMediaQuery, textWrap, zeroAprMediaQuery } from './mediaQueries';
 import { textLogoMutations, flexLogoMutations } from './common';
 
 export default {
@@ -26,9 +26,11 @@ export default {
                         width: { smallLogo: textSize * 5, largeLogo: textSize * 9 },
                         whiteSpaceBP: textSize * 27
                     }),
+                    zeroAprMediaQuery(textSize * 16),
                     `
-                    @media (max-width: ${textSize * 17}px) {
+                    @media (max-width: ${textSize * 11}px) {
                         .message__messaging { display: block; }
+                        .message__headline > .tag--small > .br:nth-child(2) { display: block; }
                     }
                     `
                 ]
@@ -37,7 +39,15 @@ export default {
         [
             'logo.type:primary && logo.position:top',
             ({ textSize }) => ({
-                styles: [`.message__logo-container { width: ${textSize * 9}px }`]
+                styles: [
+                    zeroAprMediaQuery(textSize * 16),
+                    `
+                    @media (max-width: ${textSize * 16}px) {
+                        .message__headline > .tag--small > .br:nth-child(2) { display: block; }
+                    }
+                    .message__logo-container { width: ${textSize * 9}px }
+                    `
+                ]
             })
         ],
         [
@@ -45,6 +55,7 @@ export default {
             ({ textSize }) => ({
                 messageWidth: [textSize * 10, 1000],
                 styles: [
+                    zeroAprMediaQuery(textSize * 16),
                     `
                     .message__logo-container { width: ${textSize * 9}px }
                     .message__content { display: inline-block; }
@@ -56,7 +67,7 @@ export default {
         [
             'logo.type:inline',
             ({ textSize }) => ({
-                styles: [`.message__logo { width: ${textSize * 7}px }`],
+                styles: [zeroAprMediaQuery(textSize * 16), `.message__logo { width: ${textSize * 7}px }`],
                 messageWidth: false,
                 logo: Logo.SINGLE_LINE_NO_PP.COLOR,
                 headline: {
@@ -67,19 +78,21 @@ export default {
         ],
         [
             'logo.type:none',
-            {
+            ({ textSize }) => ({
+                styles: [zeroAprMediaQuery(textSize * 16)],
                 messageWidth: false,
                 logo: false,
                 headline: {
                     replace: [['APR.', 'APR']],
                     br: ['APR']
                 }
-            }
+            })
         ],
         [
             'logo.type:alternative',
             ({ textSize }) => ({
                 styles: [
+                    zeroAprMediaQuery(textSize * 16),
                     `.message__logo-container { width: ${textSize * 5}px }`,
                     textWrap(textSize * 33, textSize, 'US')
                 ],
@@ -90,13 +103,14 @@ export default {
         [
             'logo.type:alternative && logo.position:top',
             ({ textSize }) => ({
-                styles: [`.message__logo-container { width: ${textSize * 5}px }`]
+                styles: [zeroAprMediaQuery(textSize * 16), `.message__logo-container { width: ${textSize * 5}px }`]
             })
         ],
         [
             'logo.type:alternative && logo.position:right',
             ({ textSize }) => ({
                 styles: [
+                    zeroAprMediaQuery(textSize * 16),
                     altContentMediaQuery(textSize * 34.3),
                     `.message__logo-container { width: ${textSize * 5}px }`,
                     textWrap(textSize * 33, textSize, 'US')
@@ -118,7 +132,7 @@ export default {
         [
             'ratio:1x4',
             {
-                headline: { br: ['payments', 'mo'] }
+                headline: { br: ['payments', 'mo.'] }
             }
         ],
         [