+++ packages/renderer/src/message/index.jsx
@@ -9,6 +9,10 @@ import MutatedText from './parts/MutatedText';
 import Styles from './parts/Styles';
 import CustomMessage from './parts/CustomMessage';
 
+const logMessage = val => (val !== 'string' ? JSON.stringify(val) : val);
+const logInfo = (addLog, val) => addLog('info', 'render_markup', `${logMessage(val)}`);
+const logError = (addLog, val) => addLog('error', 'render_markup', `${logMessage(val)}`);
+
 /**
  * Get all applicable rules based on user flattened options
  * and available rules to cascade
@@ -36,8 +40,100 @@ const applyCascade = curry((style, flattened, type, rules) =>
         type === Array ? [] : {}
     )
 );
+const getfontSourceRule = (addLog, fontSource) => {
+    if (!fontSource) {
+        return '';
+    }
+    const fontFormats = {
+        woff: 'woff', // woff
+        woff2: 'woff2', // woff2
+        ttf: 'ttf', // truetype
+        otf: 'otf', // opentype
+        eot: 'eot', // embedded opentype
+        svg: 'svg' // svg
+    };
+    const urlPattern = RegExp(`^(https?://.+?.(${[...Object.values(fontFormats)].join('|')}))$`);
+    const payload = {
+        fontSource,
+        validValues: []
+    };
+    try {
+        const ruleVal = (Array.isArray(fontSource) ? fontSource : [fontSource])
+            .map(url => {
+                const extension = urlPattern.exec(url)?.[2];
+                const format = fontFormats[extension];
+                if (format) {
+                    payload.validValues.push(url);
+                    return `url('${url}') format('${format}')`;
+                }
+                return '';
+            })
+            .filter(Boolean)
+            .join(', ');
+        logInfo(addLog, payload);
+        return ruleVal ? `src: ${ruleVal};` : '';
+    } catch (err) {
+        payload.error = err.stack;
+        logError(addLog, payload);
+        return '';
+    }
+};
+const getFontFamilyRule = (addLog, val) => {
+    if (!val) {
+        return '';
+    }
+    const payload = {
+        fontFamily: val,
+        validValue: null
+    };
+    const genericFamilies = {
+        // using a generic family requires the value not be quoted
+        serif: 'serif',
+        'sans-serif': 'sans-serif',
+        monospace: 'monospace',
+        cursive: 'cursive',
+        fantasy: 'fantasy',
+        'system-ui': 'system-ui',
+        'ui-serif': 'ui-serif',
+        'ui-sans-serif': 'ui-sans-serif',
+        'ui-monospace': 'ui-monospace'
+    };
+    const fontFamily = genericFamilies[val] ?? `'${val}'`;
+    if (fontFamily) {
+        payload.validValue = fontFamily;
+        logInfo(addLog, payload);
+        return `font-family: ${fontFamily}, PayPal-Sans-Big, PayPal-Sans, Arial, sans-serif; `;
+    }
+    payload.validValue = '';
+    logInfo(addLog, payload);
+    return '';
+};
+const getFontRules = (addLog, style) => {
+    const rules = [];
+    const textSize = style.layout === 'flex' ? undefined : style?.text?.size;
+    const fontSource = style?.text?.fontSource;
+    const fontFamily = fontSource ? 'MerchantCustomFont' : style?.text?.fontFamily;
 
-export default ({ options, markup, locale }) => {
+    const fontSelector = [
+        '.message__messaging', // text layout
+        '.message__messaging .message__headline span', // flex layout
+        '.message__messaging .message__sub-headline span',
+        '.message__messaging .message__disclaimer span'
+    ].join(',\n');
+
+    const textSizeRule = Number.isNaN(textSize) ? '' : `font-size: ${textSize}px; `;
+    const fontFamilyRule = getFontFamilyRule(addLog, fontFamily);
+    const fontSourceRule = getfontSourceRule(addLog, fontSource);
+
+    if (fontSource) {
+        rules.push(`@font-face {font-family: '${fontFamily}'; ${fontSourceRule}}`);
+    }
+    if (fontFamilyRule || textSizeRule) {
+        rules.push(`${fontSelector}{ ${fontFamilyRule}${textSizeRule} }`);
+    }
+    return rules;
+};
+export default ({ addLog, options, markup, locale }) => {
     const offerType = markup?.meta?.offerType;
     const style =
         options.style.layout === 'text' && options.style.preset === 'smallest'
@@ -62,14 +158,8 @@ export default ({ options, markup, locale }) => {
         rule.replace(/\.message/g, `.${localeClass} .message`)
     );
     const mutationStyleRules = mutationRules.styles ?? [];
+    const customFontStyleRules = getFontRules(addLog, style);
     const miscStyleRules = [];
-
-    const textSize = style.text?.size;
-    if (layout === 'text' && textSize) {
-        // miscStyleRules.push(`.message__headline { font-size: ${textSize}px; }`);
-        // miscStyleRules.push(`.message__disclaimer { font-size: ${textSize}px; }`);
-        miscStyleRules.push(`.message__messaging { font-size: ${textSize}px; }`);
-    }
 
     // Set boundaries on the width of the message text to ensure proper line counts
     if (mutationRules.messageWidth) {
@@ -103,6 +193,7 @@ export default ({ options, markup, locale }) => {
                     localeStyleRules={localeStyleRules}
                     mutationStyleRules={mutationStyleRules}
                     miscStyleRules={miscStyleRules}
+                    customFontStyleRules={customFontStyleRules}
                 />
             </CustomMessage>
         );
@@ -115,6 +206,7 @@ export default ({ options, markup, locale }) => {
                 localeStyleRules={localeStyleRules}
                 mutationStyleRules={mutationStyleRules}
                 miscStyleRules={miscStyleRules}
+                customFontStyleRules={customFontStyleRules}
             />
             <div className={`message__container ${localeClass}`}>
                 {/* foreground layer */}