+++ packages/renderer/tests/unit/spec/message/index.test.js
@@ -4,6 +4,9 @@ import { render } from '@testing-library/preact';
 
 import Message from '@renderer/message';
 import { getMutations, getLocaleStyles } from '@renderer/locale';
+import { objectSet } from '@library/common';
+
+const mockLogger = jest.fn();
 
 jest.mock('@paypal/sdk-logos/src', () => {
     const mock = () => ({
@@ -38,7 +41,9 @@ jest.mock('@renderer/message/styles/fonts.css', () => ({
 
 jest.mock('@renderer/message/styles', () => ({
     default: {
-        'layout:text': [['default', '']]
+        'layout:text': [['default', '']],
+        'layout:flex': [['default', '']],
+        'layout:custom': [['default', '']]
     },
     __esModule: true
 }));
@@ -88,34 +93,178 @@ describe('SSR message', () => {
 
         getLocaleStyles.mockReturnValue([]);
     });
-
-    test('renders message content', () => {
-        const logoSrc = 'logoSrc';
-
-        getMutations.mockReturnValue(
-            defaultMutations({
-                subHeadline: 'default',
-                disclaimer: 'default',
+    describe('renders message', () => {
+        const renderOptions = {
+            style: {
+                layout: 'text',
                 logo: {
-                    src: logoSrc,
-                    dimensions: [10, 10]
+                    type: 'primary',
+                    position: 'left'
+                },
+                text: {
+                    size: 12
                 }
-            })
-        );
+            }
+        };
+        test('content', () => {
+            const logoSrc = 'logoSrc';
 
-        const { getByText, getByAltText, container } = render(
-            <Message locale="US" options={options} markup={defaultMarkup()} />
-        );
+            getMutations.mockReturnValue(
+                defaultMutations({
+                    subHeadline: 'default',
+                    disclaimer: 'default',
+                    logo: {
+                        src: logoSrc,
+                        dimensions: [10, 10]
+                    }
+                })
+            );
+            const { getByText, getByAltText } = render(
+                <Message locale="US" addLog={mockLogger} options={renderOptions} markup={defaultMarkup()} />
+            );
+            expect(getByText(headline)).toBeInTheDocument();
+            expect(getByText(subHeadline)).toBeInTheDocument();
+            expect(getByText(disclaimer)).toBeInTheDocument();
 
-        expect(getByText(headline)).toBeInTheDocument();
-        expect(getByText(subHeadline)).toBeInTheDocument();
-        expect(getByText(disclaimer)).toBeInTheDocument();
+            expect(getByAltText('PayPal Credit logo')).toHaveAttribute('src', logoSrc);
+        });
+        const getMatchPattern = (cssSelector, cssValue) => {
+            // convert plain string css into an array if RegExps
+            const matchCssValue =
+                cssValue
+                    ?.replace(/\s*\{\s*/, '[^\\{]*{[^\\}]*')
+                    .replace(/;\s*\}\s*/, ';[^\\}]*}')
+                    .replace(/\(/g, '\\(')
+                    .replace(/\)/g, '\\)')
+                    .replace(/ /, '\\s*') || '';
+            const matchCssRules = cssSelector
+                .split(',')
+                .map(selectorPattern =>
+                    RegExp(
+                        selectorPattern
+                            .trim()
+                            .replace(/\./g, '\\.')
+                            .replace(/\(/g, '\\(')
+                            .replace(/\)/g, '\\)')
+                            .replace(/ +/g, '[^,\\{]') + matchCssValue || ''
+                    )
+                )
+                .filter(Boolean);
+            return matchCssRules;
+        };
+        const getRenderStyles = (logger, opts, markup, layout, prop, value) => {
+            // get the css rules used when rendered with the specified options
+            if (typeof value !== 'undefined') {
+                objectSet(opts, 'style.layout', layout);
+                objectSet(opts, prop, value);
+            }
+            const { container } = render(<Message locale="US" addLog={logger} options={opts} markup={markup} />);
+            // const style = Array.from(container.querySelectorAll('style'))
+            //     .map(e => e.textContent)
+            //     .join('\n');
+            const style = container.querySelector('.styles__customFont')?.textContent;
+            return style || '';
+        };
+        const scenarios = {
+            DEFAULT: 'default value',
+            VALID: 'valid value',
+            INVALID: 'invalid value',
+            MALICIOUS: 'malicious value'
+        };
+        const flexSelector = `
+            .message__messaging .message__headline span,
+            .message__messaging .message__sub-headline span,
+            .message__messaging .message__disclaimer span`;
+        const textSelector = `.message__messaging`;
+        const fontFaceSelector = '@font-face';
+        const fontFamilyData = {
+            'default value': [undefined, null],
+            'valid value': ['Impact', "{ font-family: 'Impact', PayPal-Sans-Big, PayPal-Sans, Arial, sans-serif; }"],
+            'invalid value': [' ', null],
+            'malicious value': ["</script><script>alert('XSS Message!')</script>", null]
+        };
+        const fontSourceData = {
+            'default value': [undefined, null],
+            'valid value': [
+                ['https://fonts.com/plRP.woff', 'https://fonts.com/plRP.woff2'],
+                "{ src: url('https://fonts.com/plRP.woff') format('woff'), url('https://fonts.com/plRP.woff2') format('woff2'); }"
+            ],
+            // 'valid value': [
+            //     'https://fonts.com/plRP.woff',
+            //     "{ src: url('https://fonts.com/plRP.woff') format('woff'); }"
+            // ],
+            'invalid value': ['https://fonts.com/plRP', null],
+            'malicious value': ["</script><script>alert('XSS Message!')</script>", null]
+        };
+        describe.each([
+            [
+                'text',
+                'text.size',
+                textSelector,
+                {
+                    'default value': [undefined, '{ font-size: 12px; }'],
+                    'valid value': [14, '{ font-size: 14px; }']
+                    // 'invalid value': [48, '{ font-size: 12px; }']
+                }
+            ],
+            [
+                'flex',
+                'text.size',
+                flexSelector,
+                {
+                    'invalid value': [14, null]
+                }
+            ],
+            ['text', 'text.fontFamily', `.message__messaging`, fontFamilyData],
+            ['flex', 'text.fontFamily', flexSelector, fontFamilyData],
+            ['text', 'text.fontSource', fontFaceSelector, fontSourceData],
+            ['flex', 'text.fontSource', fontFaceSelector, fontSourceData]
+        ])('css rule %s.%s', (layout, propString, selector, testValues) => {
+            beforeEach(() => {
+                getLocaleStyles.mockReturnValue([]);
+                getMutations.mockReturnValue([]);
 
-        expect(getByAltText('PayPal Credit logo')).toHaveAttribute('src', logoSrc);
-
-        const miscStyles = Array.from(container.querySelectorAll('style')).find(el => el.className === 'styles__misc');
-
-        expect(miscStyles.textContent).toBe('.message__messaging { font-size: 12px; }');
+                renderOptions.style = {
+                    layout: 'text',
+                    logo: {
+                        type: 'primary',
+                        position: 'left'
+                    },
+                    text: {
+                        size: 12
+                    }
+                };
+            });
+            const markup = defaultMarkup();
+            const propPath = `style.${propString}`;
+            Array.from(Object.values(scenarios)).map(scenario => {
+                const values = testValues?.[scenario];
+                if (typeof values !== 'undefined') {
+                    return test(scenario, () => {
+                        const [valueIn, valueOut] = values;
+                        const matchPatterns = getMatchPattern(selector, valueOut ?? 'placeholderValue');
+                        const styleRules = getRenderStyles(
+                            mockLogger,
+                            renderOptions,
+                            markup,
+                            layout,
+                            propPath,
+                            valueIn
+                        );
+                        // expect(objectGet(renderOptions, propPath)).toBe(valueIn);
+                        expect(typeof styleRules).toBe('string');
+                        matchPatterns.map(pattern => {
+                            expect(typeof pattern?.source).toBe('string');
+                            if (valueOut) {
+                                return expect(styleRules).toMatch(pattern);
+                            }
+                            return expect(styleRules).not.toMatch(pattern);
+                        });
+                    });
+                }
+                return test.todo(scenario);
+            });
+        });
     });
 
     test('applies cascade mutations', () => {
@@ -157,7 +306,9 @@ describe('SSR message', () => {
             [xsmallSubHeadline, ['xsmall']]
         ];
 
-        const { getByText, queryByText } = render(<Message locale="US" options={options} markup={markup} />);
+        const { getByText, queryByText } = render(
+            <Message locale="US" addLog={mockLogger} options={options} markup={markup} />
+        );
 
         expect(queryByText(headline)).toBeNull();
         expect(queryByText(subHeadline)).toBeNull();
@@ -178,7 +329,9 @@ describe('SSR message', () => {
             ]
         ]);
 
-        const { container } = render(<Message locale="US" options={options} markup={defaultMarkup()} />);
+        const { container } = render(
+            <Message locale="US" addLog={mockLogger} options={options} markup={defaultMarkup()} />
+        );
 
         const miscStyles = Array.from(container.querySelectorAll('style')).find(el => el.className === 'styles__misc');
 
@@ -192,7 +345,9 @@ describe('SSR message', () => {
             })
         );
 
-        const { container } = render(<Message locale="US" options={options} markup={defaultMarkup()} />);
+        const { container } = render(
+            <Message locale="US" addLog={mockLogger} options={options} markup={defaultMarkup()} />
+        );
 
         const miscStyles = Array.from(container.querySelectorAll('style')).find(el => el.className === 'styles__misc');
 
@@ -205,7 +360,9 @@ describe('SSR message', () => {
             ['logo.type:primary', style2]
         ]);
 
-        const { container } = render(<Message locale="US" options={options} markup={defaultMarkup()} />);
+        const { container } = render(
+            <Message locale="US" addLog={mockLogger} options={options} markup={defaultMarkup()} />
+        );
 
         const localeStyles = Array.from(container.querySelectorAll('style')).find(
             el => el.className === 'styles__locale'
@@ -222,7 +379,9 @@ describe('SSR message', () => {
             })
         );
 
-        const { container } = render(<Message locale="US" options={options} markup={defaultMarkup()} />);
+        const { container } = render(
+            <Message locale="US" addLog={mockLogger} options={options} markup={defaultMarkup()} />
+        );
 
         const mutationStyles = Array.from(container.querySelectorAll('style')).find(
             el => el.className === 'styles__mutations'