+++ tests/functional/spec/utils/modalSnapshot.js
@@ -1,4 +1,5 @@
 import { configureToMatchImageSnapshot } from 'jest-image-snapshot';
+import { logScreenshot } from './logging';
 
 const toMatchImageSnapshot = configureToMatchImageSnapshot({
     failureThresholdType: 'percent',
@@ -11,8 +12,7 @@ const toMatchImageSnapshot = configureToMatchImageSnapshot({
 expect.extend({ toMatchImageSnapshot });
 
 const modalSnapshot = async (testNameParts, viewport, account) => {
-    // eslint-disable-next-line no-console
-    console.log(`Taking screenshot of [${testNameParts}] with dimensions ${JSON.stringify(viewport)}`);
+    logScreenshot({ name: testNameParts, viewport });
 
     const image = await page.screenshot(
         {
@@ -33,6 +33,8 @@ const modalSnapshot = async (testNameParts, viewport, account) => {
         locale = 'DE';
     } else if (account.includes('GBPL')) {
         locale = 'GB';
+    } else if (account.includes('FRPL')) {
+        locale = 'FR';
     }
 
     expect(image).toMatchImageSnapshot({